Alter User Card component by reintroducing moodyButton
